moved the tests for acn and nm indexing functions to test_indexing
